interim move of read timeout stream into ClientModel
